better particles, applied to all lava
