+++ Slider/Assets/Prefabs/Particles/Lava Particle Manager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 03495b82987194c44a49d53f2fbfcae0
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
