+++ Slider/Assets/Scripts/Effects/LavaParticles.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0d978bfbaa20e5449b30c8f5ffbafd41
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
